Move `ResourceHandler` into metadata
